Finishing touches

+++ FVCalc.java
@@ -1,16 +1,17 @@
 // Computes the future value of a saving investment.
 public class FVCalc {
 	public static void main(String[] args){
-		double deposit = Double.parseDouble(args[0]);
-		double percent = (double) Double.parseDouble(args[1]) / 100.0;
+		int deposit = Integer.parseInt(args[0]);
+		double percent = Double.parseDouble(args[1]) / 100.0;
 		int years = Integer.parseInt(args[2]);
-		double yield = deposit;
+		double yield = (double) deposit;
 
 		for (int i = 0; i < years; i++) {
 			yield = yield + (yield * percent);
 		}
 
 		int yieldInt = (int) yield;
+
 		System.out.println("After " + years +" years, $" + deposit + " saved at " + (percent * 100) + "% will yield $" + yieldInt);
 	}
 }
